refactor: prevent sparse indexes on ordered models
Previously, the change of one index would not imply in a change of others.
This would lead to sparse indexes. For instance, a group with indexes:
  0, 1, 2, 3, 4
If index 1 is updated to 10, it would result in:
  0, 2, 3, 4, 10

This fixes this issue by removing any possible gap and ensuring the list
always starts at 0.
Also fixes an issue where several records could all have their index set
to 0.