feat: support to format a datetime with custom value
The method `as_str` supports format datetime by app settings or default value. In most cases
it's enought, alghout we have a case where the app' datetime format ends with Z and a single endpoint must returns the properly tz. The current method always returns "Z" even for timezone which are not in UTC