fix: manually filter-out M2M soft-deleted models
When a M2M relationship has a through soft-delete model, this intermediate model is not filtered out when accessing the M2M relationship.

Django dos not support custom managers for M2M relationships, as stated in the docs.

So we will attempt to manually detect queries that goes through the M2M model, and add the is_deleted filter to it.