fix: remove mandatory conjoined in AllOfFilter
In the previous way, the value of `conjoined` was reset in the parent
class (MultipleChoiceFilter), making it mandatory to provide the value
as `True` when using the filter. Ex.: AllOfFilter(conjoined=True).
With this change, omitting this declaration, `__init__` will fill in this
value, and the exception continues to be raised when the passed value is
False
Additionally, a typo error was fixed.