feat: add proper conversion from `/search` to `list`
Even though list is capable of retrieving the args it needs to create a queryset, it cannot benefit from caching. Caching uses parameters from `query_params` to construct the cache key, but when proxifying `search` to `list`, parameters are still in request body.

To solve this, we create a new request, using the old request body as the new request query params.