feat: allow cached search on cached views
Since there are specific views that have cached-versions of themselves, we are adding cached-searchable-version of those views as well.

Because of the way views with `@action` decorators get registered by DRF, enabling new cached-searchable viewsets requires creating new classes with specific inheritance order. This is why new classes were added.

A way of avoiding this can be by simplifying cached-search inheritance, which is not being done now.